[#33] refactor : 모든 runner 의 기능은 service 로 refactor
